ChinaStatus/requirement.txt
Django==2.1.3
django-cors-headers==2.4.0
django-filter==2.0.0
django-grappelli==2.12.1
djangorestframework==3.9.0
Markdown==3.0.1
psycopg2==2.7.5
pytz==2018.7
